% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_maps.R
\name{basic_map_united_states}
\alias{basic_map_united_states}
\title{Basic United states maps}
\usage{
basic_map_united_states(
  states_data,
  map_1_choice = 3,
  map_2_choice = 4,
  map_state
)
}
\arguments{
\item{states_data}{COVID-19 data for states single column}

\item{map_1_choice}{user input}

\item{map_2_choice}{user input}

\item{map_state}{Base map data from previous function}
}
\value{
plot
}
\description{
Basic United states maps
}
