% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prepare_data.R
\name{covid_rates}
\alias{covid_rates}
\title{Daily rates per 100,000}
\usage{
covid_rates(covid_data, population)
}
\arguments{
\item{covid_data}{Data frame: daily raw count, must contain FIPS code}

\item{population}{Data frame: population data, must contain FIPS code}
}
\value{
Daily cumulative rate per 100,000
}
\description{
Get the daily cumulative rate per 100,000 at either the state or county level.
}
