% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prepare_data.R
\name{cumulative_county}
\alias{cumulative_county}
\title{Cumulative rate per 100,000 as of latest date for counties}
\usage{
cumulative_county(county_covid, county_pop)
}
\arguments{
\item{county_covid}{Data frame: daily cumulative counts, must contain FIPS}

\item{county_pop}{Data frame: county level population data, must contain FIPS}
}
\value{
Cumulative rate per 100,000 as of latest date
}
\description{
Gets the cumulative rate per 100,000 for all counts through the latest date in the data set for counties
}
