% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prepare_data.R
\name{cumulative_us}
\alias{cumulative_us}
\title{Max Cumulative US Rate}
\usage{
cumulative_us(state_covid, us_pop)
}
\arguments{
\item{state_covid}{}

\item{us_pop}{}
}
\value{

}
\description{
Max Cumulative US Rate
}
