% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prediction.R
\name{forecast_covid}
\alias{forecast_covid}
\title{Forecasting future cases and deaths}
\usage{
forecast_covid(covid_state, days_range, days_forecast, cases_deaths)
}
\arguments{
\item{covid_state}{covid data for a state, time series}

\item{days_range}{number of days used in forecast}

\item{days_forecast}{days to forecase}

\item{cases_deaths}{new cases or new deaths}
}
\value{
plot
}
\description{
Forecasting future cases and deaths
}
