% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_maps.R
\name{map_data_states}
\alias{map_data_states}
\title{Base information for states map}
\usage{
map_data_states(cumulative_states, method)
}
\arguments{
\item{cumulative_states}{states dataset}

\item{method}{Either "state" for COVID cases and deaths or "vaccinations" 
for vaccinations}
}
\value{

}
\description{
Base information for states map
}
