% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_maps.R
\name{map_united_states}
\alias{map_united_states}
\title{United states maps}
\usage{
map_united_states(states_data, map_state, method = "cases")
}
\arguments{
\item{states_data}{COVID-19 data for states}

\item{map_state}{Base map data from previous function}

\item{method}{Select method for type of map}
}
\value{

}
\description{
United states maps
}
