% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prepare_data.R
\name{sir_counties}
\alias{sir_counties}
\title{SIR Values for Counties}
\usage{
sir_counties(cumu_county, cumu_state)
}
\arguments{
\item{cumu_county}{}

\item{cumu_state}{}
}
\value{

}
\description{
SIR Values for Counties
}
