% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_sir_plots.R
\name{sir_plot}
\alias{sir_plot}
\title{SIR Plots}
\usage{
sir_plot(sir_data, outcome, pop_level)
}
\arguments{
\item{sir_data}{Data frame: SIR values for given population level}

\item{outcome}{Boolean value: Indicator for cases or deaths}

\item{pop_level}{String: Indicator for population level}
}
\value{

}
\description{
SIR Plots
}
