% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prepare_data.R
\name{sir_states}
\alias{sir_states}
\title{SIR Values for States}
\usage{
sir_states(cumu_state, cumu_us)
}
\arguments{
\item{cumu_state}{}

\item{cumu_us}{}
}
\value{

}
\description{
SIR Values for States
}
