% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_time_series_plots.R
\name{time_series_plot}
\alias{time_series_plot}
\title{Time Series Plots}
\usage{
time_series_plot(covid_data, outcome, pop_level)
}
\arguments{
\item{covid_data}{}

\item{outcome}{}

\item{pop_level}{}
}
\value{

}
\description{
Time Series Plots
}
