% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prepare_data.R
\name{us_rates}
\alias{us_rates}
\title{US Rates}
\usage{
us_rates(state_covid, us_population)
}
\arguments{
\item{state_covid}{}

\item{us_population}{}
}
\value{
US COVID-19 rates in matrix
}
\description{
US Rates
}
\examples{


}
