% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_maps.R
\name{basic_map_united_states}
\alias{basic_map_united_states}
\title{United states difference map}
\usage{
basic_map_united_states(
  states_data,
  map_1_choice = 3,
  map_2_choice = 4,
  map_state
)
}
\arguments{
\item{states_data}{COVID-19 data for states single column}

\item{map_1_choice}{Variable from Plot 1}

\item{map_2_choice}{Variable from Plot 2}

\item{map_state}{Base map data from previous function}
}
\value{

}
\description{
United states difference map
}
