% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prepare_data.R
\name{covid_rates}
\alias{covid_rates}
\title{covid_rates}
\usage{
covid_rates(covid_data, population)
}
\arguments{
\item{covid_data}{}

\item{population}{}
}
\value{
covid_rates
}
\description{
covid_rates
}
\examples{




}
