% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prepare_data.R
\name{cumulative_county}
\alias{cumulative_county}
\title{Max Cumulative County Rate}
\usage{
cumulative_county(county_covid, county_pop)
}
\arguments{
\item{county_covid}{}

\item{county_pop}{}
}
\value{

}
\description{
Max Cumulative County Rate
}
