% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prepare_data.R
\name{cumulative_state}
\alias{cumulative_state}
\title{Max Cumulative State Rate}
\usage{
cumulative_state(state_covid, state_pop)
}
\arguments{
\item{state_covid}{}

\item{state_pop}{}
}
\value{

}
\description{
Max Cumulative State Rate
}
