% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_maps.R
\name{map_data_states}
\alias{map_data_states}
\title{States map data}
\usage{
map_data_states(cumulative_states)
}
\arguments{
\item{cumulative_states}{}
}
\value{
map_state Long/Lat data for the states
}
\description{
States map data
}
\examples{

}
