% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_prepare_data.R
\name{us_rates}
\alias{us_rates}
\title{Daily cumulative rates per 100,000}
\usage{
us_rates(state_covid, us_population)
}
\arguments{
\item{state_covid}{Data frame: state level COVID-19 cumulative count information}

\item{us_population}{Data frame: United states population data}
}
\value{
Daily US rates per 100,000
}
\description{
Returns the US daily cumulative rates per 100,000 based on information
from the state level
}
